refactor: move code to useAddBookForm hook
